DocBrain AI/requirements.txt
# requirements.txt
# Core dependencies for DocBrain AI

# Web framework
fastapi==0.104.1
uvicorn==0.24.0.post1
python-multipart==0.0.6

# PDF processing
PyPDF2==3.0.1

# OCR dependencies
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.1

# AI and ML libraries
transformers==4.49.0
sentence-transformers==3.4.1
torch==2.6.0+cu126
accelerate==1.4.0
bitsandbytes==0.45.2
huggingface_hub==0.29.1

# Data processing
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.67.1

# Note: CUDA dependencies are managed by conda environment
